backend/PinnacleBackend/requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.3.1
djangorestframework==3.14.0
sqlparse==0.5.5
